% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomasa}
\alias{biomasa}
\title{biomasa de arboles}
\format{
Data con 90 registros y 8 variables
\describe{
\item{finca}{ubicación del árbol}
\item{mg}{genotipo del árbol}
\item{bio_area}{área estimada}
\item{bio_sub}{área sub}
\item{bio_total}{área total}
\item{area_foliar}{área foliar}
\item{diametro}{diametro del árbol}
\item{altura}{altura del árbol}
}
}
\usage{
  biomasa
}
\description{
  Esta data de origen desconocido
}
\examples{
  data(biomasa)
}
\keyword{datasets}
