% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarreraLuz22}
\alias{CarreraLuz22}
\title{CarreraLuz22}
\format{
  A data frame with 1922 observations on the following 7 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{sex}}{a character vector}
    \item{\code{edad}}{a numeric vector}
    \item{\code{origen}}{a character vector}
    \item{\code{timerun}}{a numeric vector}
    \item{\code{categoria}}{a character vector}
    \item{\code{nacionalidad}}{a character vector}
  }
}
\details{
base de datos on informacion de las caracteristicas de los atletas participantes en la carrera La Lz 2022 en Cali
}
\source{
datos recogidos de la web de la carrera
}
\references{
datos recogidos de la web de la carrera
}
\examples{
data(CarreraLuz22)
## maybe str(CarreraLuz22) ; plot(CarreraLuz22) ...
}
\keyword{datasets}
