% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotacionNA}
\alias{rotacionNA}
\title{rotacionNA}
\description{
   Esta base de datos contiene los datos de 1470 empleados }
\format{
  data.frame: A data frame with 1470 observations on the following 25 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{Rotación}}{a character vector}
    \item{\code{Edad}}{a numeric vector}
    \item{\samp{Viaje de Negocios}}{a character vector}
    \item{\code{Departamento}}{a character vector}
    \item{\code{Distancia_Casa}}{a numeric vector}
    \item{\code{Educación}}{a numeric vector}
    \item{\code{Campo_Educación}}{a character vector}
    \item{\code{Satisfacción_Ambiental}}{a numeric vector}
    \item{\code{Genero}}{a character vector}
    \item{\code{Cargo}}{a character vector}
    \item{\code{Satisfación_Laboral}}{a numeric vector}
    \item{\code{Estado_Civil}}{a character vector}
    \item{\code{Ingreso_Mensual}}{a numeric vector}
    \item{\code{Trabajos_Anteriores}}{a numeric vector}
    \item{\code{Horas_Extra}}{a character vector}
    \item{\code{Porcentaje_aumento_salarial}}{a numeric vector}
    \item{\code{Rendimiento_Laboral}}{a numeric vector}
    \item{\code{Años_Experiencia}}{a numeric vector}
    \item{\code{Capacitaciones}}{a numeric vector}
    \item{\code{Equilibrio_Trabajo_Vida}}{a numeric vector}
    \item{\code{Antigüedad}}{a numeric vector}
    \item{\code{Antigüedad_Cargo}}{a numeric vector}
    \item{\code{Años_ultima_promoción}}{a numeric vector}
    \item{\code{Años_acargo_con_mismo_jefe}}{a numeric vector}
  }
}
\details{
base de datos rotacion de empleados 
}
\source{
datos sinteticos
}
\references{
datos simulados
}
\examples{
data(rotacionNA)
## maybe str(rotacionNA) ; plot(rotacionNA) ...
}
\keyword{datasets}
