% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seriesCOL}
\alias{seriesCOL}
\title{seriesCOL}
\format{
  data.frame: A data frame with 74 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{a numeric vector}
    \item{\code{INF}}{a numeric vector}
    \item{\code{TC}}{a numeric vector}
  }
}
\details{
series historicas  de la inflación y la tasa de cambio de Colombia
}
\source{
banco de republica de colombia
}
\references{
web del banco de la republica
}
\examples{
data(seriesCOL)
## maybe str(seriesCOL) ; plot(seriesCOL) ...
}
\keyword{datasets}
