% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarreraLuz22}
\alias{CarreraLuz22}
\title{Tiemos de carrera la Luz 2022}
\format{
Data con 1922 registros y 7 variables
\describe{
\item{id}{identificador del atleta.}
\item{sex}{sexo del atleta.}
\item{edad}{edad del atleta.}
\item{origen}{ciudad de origen del atleta.}
\item{timerun}{tiempo de carrera en segundos.}
\item{categoria}{categoria del atleta.}
\item{nacionalidad}{nacionalidad del atleta.}
}
}
\usage{
  CarreraLuz22
}
\description{
  Esta data contiene información de 1922 atletas que participaron en la carrera de 10.7 km en la ciudad de Cali al final del 2022.
}
\examples{
  data(CarreraLuz22)
}
\keyword{datasets}
