% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarreraLuz22}
\alias{CarreraLuz22}
\title{CarreraLuz22}
\format{
  A data frame with 1922 observations on the following 7 variables.
}
\description{
  Base de datos con información de las características de los atletas participantes en la carrera La Luz 2022 en Cali.
}
\details{
  Datos recogidos de la web de la carrera.
}
\source{
  Datos recogidos de la web de la carrera.
}
\references{
  Datos recogidos de la web de la carrera.
}
\examples{
  data(CarreraLuz22)
  ## maybe str(CarreraLuz22) ; plot(CarreraLuz22) ...
}
\keyword{datasets}
